requirements.txt
# Core dependencies
pandas>=1.3.0             # For data manipulation
openpyxl>=3.0.0           # For working with Excel files

# Box SDK for interacting with Box APIs
boxsdk>=3.0.0             # Box Python SDK

# Machine learning library
scikit-learn>=1.0.0       # For regression and other ML models

# Environment variable management
python-dotenv>=1.0.0      # For loading .env files

# Optional (if running tests or linting)
pytest>=7.0.0             # For writing and running tests
flake8>=4.0.0             # For linting and code style checks (optional)